Move to CMake, refactor project structure.
Disabled Qt keywords because it clashes with glib keywords. This is usually solved by making sure glib is included before Qt but in this project that solution is difficult to implement due to project size and its usage of mafw. So we'll use Q_SIGNALS, Q_SLOTS, Q_FOREACH, and Q_EMIT instead.